NIFI-12684: - Introducing parameter helper to remove duplicated logic and error handling when fetching parameters and converting properties to parameters.
