+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-frontend/src/main/nifi/src/app/pages/flow-designer/state/parameter/parameter.reducer.ts
@@ -45,10 +45,7 @@ export const parameterReducer = createReducer(
         ...state,
         updateRequestEntity: response.requestEntity
     })),
-    on(editParameterContextComplete, (state) => ({
-        ...state,
-        saving: false,
-        updateRequestEntity: null,
-        error: null
+    on(editParameterContextComplete, () => ({
+        ...initialState
     }))
 );
